--- Program.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Windows.Forms;
-
-namespace EncryptionHeckingCode
-{
-    static class Program
-    {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new EncryptionMainForm());
-        }
-    }
-}
