+++ code/Program.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Windows.Forms;
+
+namespace EncryptionHeckingCode
+{
+    static class Program
+    {
+        /// <summary>
+        /// The main entry point for the application.
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Application.Run(new EncryptionMainForm());
+        }
+    }
+}
