+++ code/Properties/AssemblyInfo.cs
