+++ code/Properties/Resources.resx
