+++ code/bin/Debug/Hecking encryption.vshost.exe.manifest
